State and county names by FIPS codes

A dataset that links state and county FIPS codes in the US (as character values) and province and census division codes (Canada) to the names of those states/provinces and counties/census divisions. This data predates the inclusion of Canada in SafeGraph, thus the US-centric naming.

fips_to_names

Format

A data.table with 3142 rows and 4 variables:

state_fips

State FIPS code / Canadian SGC code

county_fips

County FIPS code / Canadian Census division

statename

The full name of the state / province

countyname

The full English name of the county / census division, including "County" for US entries. Merge with canada_cd_types to get the equivalent division type for Canada and French names.

iso_country_code

Indicator for US or Canada

Source

US Census