Example Patterns Data Aggregated by NAICS Code

The patterns file from the July 1 weekly patterns pull, aggregated to the NAICS code level for the vignette. See [SafeGraph Docs](https://docs.safegraph.com/docs) for full documentation.

The patterns file from the July 1 weekly patterns pull, aggregated to the county level for the vignette, for New York and New Jersey only. See [SafeGraph Docs](https://docs.safegraph.com/docs) for full documentation.

pat_naics

pat_naics

Format

A data.table with 9247 rows and 7 variables:

date

The date

start_date,day

The first date present in the patterns file, and whether this observation is from that first date (1), the second (2), etc.

state_fips,county_fips

Originally would have been the state and county identifiers, but since things were aggregated to the NAICS level (rather than NAICS/state/county), they have been summed up and now mean nothing.

naics_code

Six-digit NAICS code

visits_by_day

The total number of visits to POIs of this NAICS code on this day

A data.table with 2324 rows and 7 variables:

date

The date

start_date,day

The first date present in the patterns file, and whether this observation is from that first date (1), the second (2), etc.

state_fips,county_fips

The state and county identifiers

naics_code

Originally this was the six-digit NAICS code of the associated POI. But since aggregation didn't preserve NAICS, this is nonsense

visits_by_day

The total number of visits to POIs in this county on this day

Source

SafeGraph

SafeGraph