Read SafeGraph Core

Be aware that the files this is designed to work with are large and this function may take a while to execute. This function takes a single .zip SafeGraph Core file and reads it in. The output is a data.table.

read_core(
  filename,
  dir = ".",
  exdir = ".",
  cleanup = TRUE,
  filter = NULL,
  select = NULL,
  silent = FALSE,
  ...
)

Arguments

filename
The filename of the ZIP Core Places file.

dir
The directory that the file is in.

exdir
Name of the directory to unzip to.

cleanup
Set to TRUE to delete all the unzipped files after being read in.

filter
A character string describing a logical statement for filtering the data, for example filter = 'naics_code == 512131' would give you only movie theater POIs. Will be used as an i argument in a data.table, see help(data.table). Filtering here instead of afterwards can cut down on time and memory demands.

select
Character vector of variables to get from the file. Set to NULL to get all variables. If you plan to link the results to a patterns file, you will probably want to include 'safegraph_place_id' or 'placekey' in this vector. Note that any variables mentioned in filter MUST be in select unless select = NULL.

silent
Suppress timing messages.

...
Other arguments to be passed to data.table::fread when reading in the CSV files inside of the ZIP. For example, nrows to only read in a certain number of rows.

Details

If you want all of the rows and only want to extract their naics_code, you can use the deprecated but still-working link_poi_naics() function.

Examples

if (FALSE) {
# Location of our CORE file
dir <- '../SafeGraph/2020/10/'
filename <- 'Core-USA-Oct-CORE_POI-2020_09-2020-10-19.zip'

# Let's only get retail POIs in California
# And
locations <- read_core(filename = filename,
                       dir = dir,
                       filter = 'region == "CA" & floor(naics_code/10000) %in% 44:45')
}